move KFold to algorithms/crossvalidation.py
